fix: change creator pattern concept for EncryptionService
* fixes also https://github.com/zalando/zappr/issues/132
